requirements.txt
flask
flask-sqlalchemy
pyotp
qrcode[pil]
cryptography